Update format.py
change modelling of hydrounits to be consistent with sms++ : turbines were modelled as 2 pieces functions, with p1<=pmax, p2<=CoeffSpillage*Pmax, because we had understood that in sms++ we had P_i=A+B*Flow_i, while inreality sms++ has this equality constraint for pumps, but for turbine it is P_i<=A+B*Flow_i. Which means that we don't need 2 pieces, but one is enough, with P<=Pmax, Flow<=(1+CoeffSpillage)Pmax ; then as P<=A+B*Flow, we have a turbine that can produce at max power Pmax, and that for higher flows can spill water.